// src/codec/tendermint/liquidity/v1beta1/tx.d.ts
export interface MsgCreatePool {
    poolCreatorAddress: string;
    /** id of the target pool type, must match the value in the pool. Only pool-type-id 1 is supported. */
    poolTypeId: number;
    /** reserve coin pair of the pool to deposit. */
    depositCoins: Coin[];
}
/** MsgCreatePoolResponse defines the Msg/CreatePool response type. */
export interface MsgCreatePoolResponse {
}
// ...
export interface MsgDepositWithinBatch {
    depositorAddress: string;
    /** id of the target pool */
    poolId: Long;
    /** reserve coin pair of the pool to deposit */
    depositCoins: Coin[];
}
/** MsgDepositWithinBatchResponse defines the Msg/DepositWithinBatch response type. */
export interface MsgDepositWithinBatchResponse {
}
// ...
export interface MsgWithdrawWithinBatch {
    withdrawerAddress: string;
    /** id of the target pool */
    poolId: Long;
    poolCoin?: Coin;
}
/** MsgWithdrawWithinBatchResponse defines the Msg/WithdrawWithinBatch response type. */
export interface MsgWithdrawWithinBatchResponse {
}
// ...
export interface MsgSwapWithinBatch {
    /** address of swap requester */
    swapRequesterAddress: string;
    /** id of swap type, must match the value in the pool. Only `swap_type_id` 1 is supported. */
    poolId: Long;
    /** id of swap type. Must match the value in the pool. */
    swapTypeId: number;
    /** offer sdk.coin for the swap request, must match the denom in the pool. */
    offerCoin?: Coin;
    /** denom of demand coin to be exchanged on the swap request, must match the denom in the pool. */
    demandCoinDenom: string;
    /** half of offer coin amount * params.swap_fee_rate and ceil for reservation to pay fees. */
    offerCoinFee?: Coin;
    /**
     * limit order price for the order, the price is the exchange ratio of X/Y
     * where X is the amount of the first coin and Y is the amount
     * of the second coin when their denoms are sorted alphabetically.
     */
    orderPrice: string;
}
/** MsgSwapWithinBatchResponse defines the Msg/Swap response type. */
export interface MsgSwapWithinBatchResponse {
}
export declare const MsgCreatePool: {
    encode(message: MsgCreatePool, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgCreatePool;
    fromJSON(object: any): MsgCreatePool;
    toJSON(message: MsgCreatePool): unknown;
    fromPartial(object: DeepPartial<MsgCreatePool>): MsgCreatePool;
};
export declare const MsgCreatePoolResponse: {
    encode(_: MsgCreatePoolResponse, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgCreatePoolResponse;
    fromJSON(_: any): MsgCreatePoolResponse;
    toJSON(_: MsgCreatePoolResponse): unknown;
    fromPartial(_: DeepPartial<MsgCreatePoolResponse>): MsgCreatePoolResponse;
};
export declare const MsgDepositWithinBatch: {
    encode(message: MsgDepositWithinBatch, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgDepositWithinBatch;
    fromJSON(object: any): MsgDepositWithinBatch;
    toJSON(message: MsgDepositWithinBatch): unknown;
    fromPartial(object: DeepPartial<MsgDepositWithinBatch>): MsgDepositWithinBatch;
};
export declare const MsgDepositWithinBatchResponse: {
    encode(_: MsgDepositWithinBatchResponse, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgDepositWithinBatchResponse;
    fromJSON(_: any): MsgDepositWithinBatchResponse;
    toJSON(_: MsgDepositWithinBatchResponse): unknown;
    fromPartial(_: DeepPartial<MsgDepositWithinBatchResponse>): MsgDepositWithinBatchResponse;
};
export declare const MsgWithdrawWithinBatch: {
    encode(message: MsgWithdrawWithinBatch, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgWithdrawWithinBatch;
    fromJSON(object: any): MsgWithdrawWithinBatch;
    toJSON(message: MsgWithdrawWithinBatch): unknown;
    fromPartial(object: DeepPartial<MsgWithdrawWithinBatch>): MsgWithdrawWithinBatch;
};
export declare const MsgWithdrawWithinBatchResponse: {
    encode(_: MsgWithdrawWithinBatchResponse, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgWithdrawWithinBatchResponse;
    fromJSON(_: any): MsgWithdrawWithinBatchResponse;
    toJSON(_: MsgWithdrawWithinBatchResponse): unknown;
    fromPartial(_: DeepPartial<MsgWithdrawWithinBatchResponse>): MsgWithdrawWithinBatchResponse;
};
export declare const MsgSwapWithinBatch: {
    encode(message: MsgSwapWithinBatch, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgSwapWithinBatch;
    fromJSON(object: any): MsgSwapWithinBatch;
    toJSON(message: MsgSwapWithinBatch): unknown;
    fromPartial(object: DeepPartial<MsgSwapWithinBatch>): MsgSwapWithinBatch;
};
export declare const MsgSwapWithinBatchResponse: {
    encode(_: MsgSwapWithinBatchResponse, writer?: _m0.Writer): _m0.Writer;
    decode(input: _m0.Reader | Uint8Array, length?: number | undefined): MsgSwapWithinBatchResponse;
    fromJSON(_: any): MsgSwapWithinBatchResponse;
    toJSON(_: MsgSwapWithinBatchResponse): unknown;
    fromPartial(_: DeepPartial<MsgSwapWithinBatchResponse>): MsgSwapWithinBatchResponse;
};
// ...
declare type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined | Long;
export declare type DeepPartial<T> = T extends Builtin ? T : T extends Array<infer U> ? Array<DeepPartial<U>> : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>> : T extends {} ? {
    [K in keyof T]?: DeepPartial<T[K]>;
} : Partial<T>;